+++ postfix/src/util/mvect.h
@@ -14,18 +14,18 @@
  /*
   * Generic memory vector interface.
   */
-typedef void (*MVECT_FN) (char *, int);
+typedef void (*MVECT_FN) (char *, ssize_t);
 
 typedef struct {
     char   *ptr;
-    int     elsize;
-    int     nelm;
+    ssize_t elsize;
+    ssize_t nelm;
     MVECT_FN init_fn;
     MVECT_FN wipe_fn;
 } MVECT;
 
-extern char *mvect_alloc(MVECT *, int, int, MVECT_FN, MVECT_FN);
-extern char *mvect_realloc(MVECT *, int);
+extern char *mvect_alloc(MVECT *, ssize_t, ssize_t, MVECT_FN, MVECT_FN);
+extern char *mvect_realloc(MVECT *, ssize_t);
 extern char *mvect_free(MVECT *);
 
 /* LICENSE
